+++ src/Plutarch/Test/QuickCheck/Internal.hs
@@ -19,6 +19,7 @@ module Plutarch.Test.QuickCheck.Internal (
 
 import Data.ByteString (ByteString)
 import qualified Data.Text as T (intercalate, pack, unpack)
+import Data.Word (Word8)
 import qualified GHC.Exts as Exts (IsList (fromList, toList))
 import Plutarch (
   Config (Config, tracingMode),
@@ -58,9 +59,12 @@ import Plutarch.Api.V2 (
   AmountGuarantees (NoGuarantees),
   KeyGuarantees (Unsorted),
   PAddress (PAddress),
+  PDatumHash (PDatumHash),
   PMaybeData (PDJust, PDNothing),
   PPubKeyHash (PPubKeyHash),
   PStakingCredential (PStakingHash, PStakingPtr),
+  PTxId (PTxId),
+  PTxOutRef (PTxOutRef),
  )
 import Plutarch.Evaluate (evalScriptHuge, evalTerm)
 import Plutarch.Extra.Maybe (
@@ -70,6 +74,7 @@ import Plutarch.Extra.Maybe (
  )
 import Plutarch.Lift (PUnsafeLiftDecl (PLifted), plift, plift')
 import Plutarch.Maybe (pfromJust)
+import Plutarch.Num (pabs)
 import Plutarch.Positive (PPositive, ptryPositive)
 import Plutarch.Prelude (
   PAsData,
@@ -112,10 +117,12 @@ import Test.QuickCheck (
   Gen,
   Testable (property),
   chooseInt,
+  chooseInteger,
   counterexample,
   elements,
   frequency,
   functionMap,
+  listOf,
   sized,
   variant,
   vectorOf,
@@ -604,12 +611,11 @@ instance
     where
       unMap = flip pmatchT $ \(PMap a) -> a
 
--- | @since 2.0.0
+-- | @since 2.2.2
 instance PArbitrary PPOSIXTime where
   parbitrary = do
-    (TestableTerm x) <- parbitrary
-    return $ pconT $ PPOSIXTime x
-
+    TestableTerm x <- parbitrary
+    return . pconT $ PPOSIXTime $ pabs # x
   pshrink = fmap (\(TestableTerm x) -> pconT $ PPOSIXTime x) . shrink . unTime
     where
       unTime = flip pmatchT $ \(PPOSIXTime a) -> a
@@ -670,6 +676,13 @@ instance
         PInterval $
           pdcons @"from" # pdata lo #$ pdcons @"to" # pdata up # pdnil
 
+-- | @since 2.2.2
+instance PArbitrary PDatumHash where
+  parbitrary = do
+    -- PDatumHash should be 32 bytes long
+    bs <- genByteString 32
+    return . pconT $ PDatumHash $ pconstant bs
+
 -- | @since 2.0.0
 instance PArbitrary PPubKeyHash where
   parbitrary = do
@@ -701,13 +714,36 @@ instance PArbitrary PCredential where
       , pconT $ PPubKeyCredential $ pdcons @"_0" # pdata pk # pdnil
       ]
 
--- | @since 2.0.0
+-- | @since 2.2.2
+instance PArbitrary PTxId where
+  parbitrary = do
+    -- PTxId should be 32 bytes long
+    bs <- genByteString 32
+    pure . pconT $ PTxId $ pdcons @"_0" # pdata (pconstant bs) # pdnil
+
+-- | @since 2.2.2
+instance PArbitrary PTxOutRef where
+  parbitrary = do
+    TestableTerm id' <- parbitrary
+    TestableTerm idx <- abs <$> parbitrary
+    pure . pconT $ PTxOutRef $ pdcons # pdata id' #$ pdcons # pdata idx # pdnil
+  pshrink (TestableTerm ref) = do
+    TestableTerm idx' <- pshrink $ TestableTerm $ pfield @"idx" # ref
+    pure . pconT $
+      PTxOutRef $
+        pdcons
+          # (pfield @"id" # ref)
+          #$ pdcons
+          # idx'
+          # pdnil
+
+-- | @since 2.2.2
 instance PArbitrary PStakingCredential where
   parbitrary = do
     (TestableTerm cred) <- parbitrary
-    (TestableTerm x) <- parbitrary
-    (TestableTerm y) <- parbitrary
-    (TestableTerm z) <- parbitrary
+    (TestableTerm x) <- go
+    (TestableTerm y) <- go
+    (TestableTerm z) <- go
     elements
       [ pconT $ PStakingHash $ pdcons @"_0" # pdata cred # pdnil
       , pconT $
@@ -720,6 +756,11 @@ instance PArbitrary PStakingCredential where
               # pdata z
               # pdnil
       ]
+    where
+      -- This is needed because StakingPtr indexes are meant to be bounded to the
+      -- Word64 range.
+      go :: Gen (TestableTerm PInteger)
+      go = pconstantT <$> chooseInteger (0, 18_446_744_073_709_551_615)
 
 -- | @since 2.0.0
 instance PArbitrary PAddress where
@@ -741,12 +782,26 @@ instance PArbitrary PCurrencySymbol where
     (TestableTerm cs) <- parbitrary
     return $ pconT $ PCurrencySymbol cs
 
--- | @since 2.0.0
+{- | This generates only those 'PTokenName's that correspond to ASCII strings.
+ This is somewhat limited, but otherwise would require UTF-8 encoding as part
+ of the generator.
+
+ Unlike the equivalent Haskell type generator, this instance does not shrink.
+ While somewhat suboptimal, this would require either a lot of lifting or
+ \'lowering\' back into Haskell for the shrink.
+
+ @since 2.2.2
+-}
 instance PArbitrary PTokenName where
   parbitrary = do
-    len <- chooseInt (1, 32)
-    tn <- genByteString len
-    return $ pconT $ PTokenName $ pconstant tn
+    asList <- listOf genAsciiByte
+    let bs = Exts.fromList asList
+    pure . pconT $ PTokenName $ pconstant bs
+    where
+      -- This generates only those bytes which are readable (so, not control
+      -- sequences)
+      genAsciiByte :: Gen Word8
+      genAsciiByte = fromIntegral <$> chooseInt (32, 126)
 
 -- | @since 2.0.0
 instance PArbitrary (PValue 'Unsorted 'NoGuarantees) where